This creates Rotating wheel to set temperature.
You have to implement the bottom wheel control. It needs to change the
value of the large text control marked with value 78. When the user
swipes over the wheel the value changes and as the values change the
color of the wheel also changes with it being full blue at 50 and full
red at 100. Mid point 75 is green and rest of the values are a gradient
of the blue-green or green-red for 51-74 and 76-99 range.